Fix Issue: Violation of Usage of Android Advertising ID policy and section 4.8 of the Developer Distribution Agreement
The app was removed from Google Play and I am updating Crashlytics SDK
as recommend by Fabric:

"Mike from Fabric here. I've tested this in multiple different ways and
can confirm I'm not seeing the Android Ad Id submitted on recent
versions. If you've updated to a newer version of the Crashlytics SDK,
make sure you run: ./gradlew build --refresh-dependencies otherwise it's
possible you've update the Crashlytics SDK, but not the root Fabric SDK
that collected these values.

To clarify my previous comment: From Crashlytics SDK version 2.9.3 and
higher, we no longer submit the Android Ad Id to our servers. In prior
versions of the Crashlytics SDK, this Id was primarily collected and
submitted for Mobile App Conversion Tracking and Audience Insights (both
of which are now deprecated features).

Starting with 2.9.3, we only monitor the Android Ad Id to see if it
changes in order to update our 64 bit UUID.

This UUID is used to identify which devices have experienced a crash. If
a user resets their Android Ad Id, we reset our UUID as well. If the
user deletes and re-installs the app, we also reset our UUID.

You will see our SDK reference the value, but rest assured that we never
transmit the Android Ad Id off of the device starting with version 2.9.3
of our Android SDK."